separated filters logic into hook
