+++ web/src/data/static/token-list-filters.tsx
@@ -1,0 +1,14 @@
+export const tokenListSort = [
+  { id: 1, name: 'Hot' },
+  { id: 2, name: 'APR' },
+  { id: 3, name: 'Earned' },
+  { id: 4, name: 'Total staked' },
+  { id: 5, name: 'Latest' },
+];
+
+export const tokenListStatuses = [
+  { id: 1, name: 'all' },
+  { id: 2, name: 'hideMicro' },
+];
+
+export const microDropValue = 10000;
